Covid-detection-using-chest-X-Ray/requirements.txt
keras==2.8.0
numpy==1.22.3
Pillow==9.1.1
prediction==0.0.3
streamlit==1.9.0
tensorflow==2.8.0
